progress bar UX
add a display for what time is moused over, and fix a couple minor issues with the progress bar

+++ Assets/__Scripts/UI/Text/TimeText.cs
@@ -1,4 +1,3 @@
-using System;
 using UnityEngine;
 using TMPro;
 
@@ -10,14 +9,14 @@ public class TimeText : MonoBehaviour
 
     public void UpdateText(float beat)
     {
-        uint currentTime = (uint)TimeManager.CurrentTime;
-        uint currentSeconds = currentTime % 60;
-        uint currentMinutes = currentTime / 60;
+        ulong currentTime = (ulong)TimeManager.CurrentTime;
+        ulong currentSeconds = currentTime % 60;
+        ulong currentMinutes = currentTime / 60;
 
         string secondsString = currentSeconds >= 10 ? $"{currentSeconds}" : $"0{currentSeconds}";
 
         timeText.text = $"{currentMinutes}:{secondsString}";
-        beatText.text = ((uint)beat).ToString();
+        beatText.text = ((ulong)beat).ToString();
     }
